requirements.txt
streamlit 
langchain 
langchain-google-genai
selenium
beautifulsoup4
lxml 
html5lib
python-dotenv